dynamically redefine route helper methods
because a parent app's layout file can be used, and that layout
file may already use _url and _path helper methods, for example
in a navigation menu, the engine app has to support those methods
dynamically. this code enables those methods to be redefined within
the scope that they are needed within the engine, and simply delegates
them back up to the `main_app` which should preserve their behavior.

i don't believe this is the ideal place for these methods to exist
but i didn't find a way to make them work when defined in a helper
or elsewhere yet.